feat/issue_1: Refatoração das interfaces, visando organização e manutenibilidade
Para maior organização do código, seguindo boas práticas da OO, dividi as telas principais em diferentes classes que cumprem papéis específicos para a construção da interface
