+++ Views/UsuarioMenuView.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SistemaGestaoProjetosETarefas.Views
+{
+    public class UsuarioMenuView
+    {
+        public static void MenuUsuario()
+        {
+
+        }
+    }
+}
